LEXEVS-4698. Some running code in place for the complete API. Beta status with some questions to be answered yet

+++ lbImpl/src/org/LexGrid/LexBIG/Utility/ServiceUtility.java
@@ -760,7 +760,7 @@ public class ServiceUtility {
                         ref.getCodingSchemeURN(), 
                         ref.getCodingSchemeVersion(), 
                         associationName, 
-                        entityCode) > 0;
+                        entityCode).intValue() > 0;
     }
